COLDWEB-9 Метод process вынесен в отдельный класс Process  и объеденен с интерфейсом Action
