--- src/main/java/com/github/muhin007/coldplaceweb/Action.java
@@ -1,10 +1,0 @@
-package com.github.muhin007.coldplaceweb;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public interface Action {
-    void action(HttpServletRequest request, HttpServletResponse response)throws IOException;
-
-}
