+++ src/main/java/com/github/muhin007/coldplaceweb/Servlets/URLColdplaceServlet.java
@@ -1,9 +1,9 @@
 package com.github.muhin007.coldplaceweb.Servlets;
 
-import com.github.muhin007.coldplaceweb.Action;
 import com.github.muhin007.coldplaceweb.Data.City;
 import com.github.muhin007.coldplaceweb.Data.DBConnection;
 import com.github.muhin007.coldplaceweb.PageGenerator;
+import com.github.muhin007.coldplaceweb.Process;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -18,20 +18,11 @@ import java.util.Map;
 
 public class URLColdplaceServlet extends HttpServlet {
 
-    private static void process(HttpServletRequest request,
-                                HttpServletResponse response, Action action) throws IOException {
-        response.setCharacterEncoding("UTF-8");
-        response.setContentType("text/html;charset=utf-8");
-
-        action.action(request, response);
-
-        response.setStatus(HttpServletResponse.SC_OK);
-    }
 
     public void doGet(HttpServletRequest request,
                       HttpServletResponse response) throws IOException {
 
-        URLColdplaceServlet.process(request, response, (HttpServletRequest req, HttpServletResponse resp) -> {
+        Process.process(request, response, (HttpServletRequest req, HttpServletResponse resp) -> {
                     Map<String, Object> pageVariables = createPageVariablesMap(req);
                     pageVariables.put("cityName", "");
                     resp.getWriter().println(PageGenerator.instance().getPage("URLReadPage.html", pageVariables));
@@ -42,7 +33,7 @@ public class URLColdplaceServlet extends HttpServlet {
 
     public void doPost(HttpServletRequest request,
                        HttpServletResponse response) throws IOException {
-        URLColdplaceServlet.process(request, response, (HttpServletRequest req, HttpServletResponse resp) -> {
+        Process.process(request, response, (HttpServletRequest req, HttpServletResponse resp) -> {
                     Map<String, Object> pageVariables = createPageVariablesMap(req);
 
                     String message = req.getParameter("cityName");
@@ -85,5 +76,6 @@ public class URLColdplaceServlet extends HttpServlet {
         pageVariables.put("parameters", request.getParameterMap().toString());
         return pageVariables;
     }
+
 }
